服务器协议

+++ LuTu/Classes/Models/ProtocolModels/GenModels/PMCategoryListRsp.h
@@ -1,10 +1,12 @@
 
 #import <Mantle.h>
+#import "PMCategoryInfo.h" 
+
 
 @interface PMCategoryListRsp : MTLModel <MTLJSONSerializing>
-@property(nonatomic,assign) int32_t offset;
-@property(nonatomic,assign) int32_t limit;
-@property(nonatomic,assign) int32_t total;
+@property(nonatomic,assign) int64_t offset;
+@property(nonatomic,assign) int64_t limit;
+@property(nonatomic,assign) int64_t total;
 @property(nonatomic,strong) NSArray* list;
 
 @end
